Separating callSheets and normalizeHeaders functions into their own file (#62)
